RANCHER-1365. The TestResult and TestType classes were moved from the shared folder to the testing folder. The shared folder has been deleted.
