+++ src/org/folio/testing/TestResult.groovy
@@ -1,4 +1,4 @@
-package org.folio.shared
+package org.folio.testing
 
 enum TestResult{
   SUCCESS, FAILURE
