+++ vars/karateFlow.groovy
@@ -1,7 +1,7 @@
 import groovy.json.JsonOutput
 import groovy.text.StreamingTemplateEngine
 import org.folio.Constants
-import org.folio.shared.TestType
+import org.folio.testing.TestType
 import org.folio.utilities.RestClient
 import org.jenkinsci.plugins.workflow.libs.Library
